Add encodings
git-svn-id: file:///.data/svn_convert/svn@204 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ trunk/sim/userbox.cpp
@@ -216,7 +216,7 @@ void UserBox::wmChanged()
 
 void UserBox::showEncodingPopup()
 {
-    for (int i = 0; i < menuEncoding->count(); i++)
+    for (unsigned i = 0; i < (unsigned)(menuEncoding->count()); i++)
         menuEncoding->setItemChecked(menuEncoding->idAt(i), false);
     if (curWnd){
         int encoding = pClient->userEncoding(curWnd->Uin());
